backend/requirements.txt
Django
djangorestframework
celery
redis
psycopg2-binary
gunicorn
httpx
python-dotenv
dj-database-url
whitenoise
djangorestframework
djangorestframework-simplejwt
celery
redis
httpx
